Release Telux_master_v1.38.0
Add support for getting eUICC identifier and to provide user consent.
Improve Cv2x Daemon data call KPI
Added APIs to switch active slot in Dual-Sim Single Active platforms
Adding support for Audio in telsdk_console_app
Added fix for the hardcorded NW interface of cv2x_tcp_app
Rename Data filter config parser to DataConfigParser
Refactor Data Menu
Hotfix: Fix Data Call Statistics

CRs-Fixed: 2729171, 2741543, 2729307, 2713127, 2718490, 2733463, 2714847, 2727860, 2750620

+++ include/telux/data/DataDefines.hpp
@@ -141,12 +141,12 @@ struct ProfileParams {
  * Data transfer statistics structure.
  */
 struct DataCallStats {
-    unsigned long packetsTx = 0;        /**< Number of packets transmitted */
-    unsigned long packetsRx = 0;        /**< Number of packets received */
-    long long bytesTx = 0;              /**< Number of bytes transmitted */
-    long long bytesRx = 0;              /**< Number of bytes received */
-    unsigned long packetsDroppedTx = 0; /**< Number of transmit packets dropped */
-    unsigned long packetsDroppedRx = 0; /**< Number of receive packets dropped */
+    uint64_t packetsTx = 0;               /**< Number of packets transmitted */
+    uint64_t packetsRx = 0;               /**< Number of packets received */
+    uint64_t bytesTx = 0;                 /**< Number of bytes transmitted */
+    uint64_t bytesRx = 0;                 /**< Number of bytes received */
+    uint64_t packetsDroppedTx = 0;        /**< Number of transmit packets dropped */
+    uint64_t packetsDroppedRx = 0;        /**< Number of receive packets dropped */
 };
 
 /**
